Feat (catalog)/build domain layer

+++ src/Authorization/Authorization.Application/UseCases/V1/Commands/LogoutCommandHandler.cs
@@ -20,12 +20,28 @@ public class LogoutCommandHandler : ICommandHandler<Command.LogoutCommand>
 
     public async Task<Result> Handle(Command.LogoutCommand request, CancellationToken cancellationToken)
     {
+        /*
+            1. Get Key-Value by email from Redis
+            2. Validate token user's request => verify token
+            3. Compare email from token with email from request
+            4. Remove key-value from Redis
+         */
+
+        // 1.
         var authValue = await _cacheService.GetAsync<RedisKeyValue.AuthenticatedValue>(request.Email, cancellationToken)
             ?? throw new IdentityException.TokenException("Can not get value from Redis");
 
+        // 2.
         var principle = _jwtTokenService.GetPrincipalFromExpiredToken(request.AccessToken, authValue.PublicKey);
         var emailKey = principle.FindFirstValue(ClaimTypes.Email).ToString();
 
+        // 3.
+        if (emailKey != request.Email)
+        {
+            throw new IdentityException.TokenException("Email from token is not match with email from request");
+        }
+
+        // 4.
         await _cacheService.RemoveAsync(emailKey, cancellationToken);
 
         return Result.Success();
